`browse`: move the new `_ListPromptWithCustomKeys` class to its own file
this keeps the main `browse.py` file more focused.

thanks to [at]kayoub5 for bringing this up.
